feat: DetailBox 공통 컴포넌트 UI, 애니메이션 구현(#152)
